refactor: split validators logic, remove validate_phone_number from extra_kwargs in CustomerSerializer.Meta
