+++ backend/backend/apps/customers/serializers.py
@@ -2,9 +2,8 @@ import phonenumbers
 from django.utils.translation import gettext_lazy as _
 from rest_framework import serializers
 
-from backend.apps.customers import validators
 from backend.apps.customers.models import Customer
-from backend.utils.validators import is_valid_phone_number
+from backend.utils.validation import is_valid_phone_number
 
 
 class CustomerSerializer(serializers.ModelSerializer):
@@ -13,9 +12,7 @@ class CustomerSerializer(serializers.ModelSerializer):
         fields = ["id", "nickname", "phone_number"]
         extra_kwargs = {
             "phone_number": {
-                "validators": [
-                    validators.validate_phone_number
-                ],  # excluding the UniqueValidator
+                "validators": [],  # excluding the UniqueValidator
             },
         }
         read_only_fields = ["id"]
