+++ backend/backend/utils/django_validators.py
@@ -1,7 +1,7 @@
 from django.core.exceptions import ValidationError as DjangoValidationError
 from django.utils.translation import gettext_lazy as _
 
-from backend.utils.validators import is_valid_phone_number
+from backend.utils.validation import is_valid_phone_number
 
 
 def validate_phone_number(value: str) -> None:
